+++ src/Container/AuthorizationServiceFactory.php
@@ -16,26 +16,41 @@
  * and is licensed under the MIT license.
  */
 
-namespace ZfcRbac\Factory;
+namespace ZfcRbac\Container;
 
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
-use ZfcRbac\Options\ModuleOptions;
+use ZfcRbac\Service\AuthorizationService;
 
 /**
- * Factory for the module options
+ * Factory to create the authorization service
  *
  * @author  Michaël Gallego <mic.gallego@gmail.com>
  * @licence MIT
  */
-class ModuleOptionsFactory implements FactoryInterface
+class AuthorizationServiceFactory implements FactoryInterface
 {
     /**
      * {@inheritDoc}
-     * @return ModuleOptions
+     * @return AuthorizationService
      */
     public function createService(ServiceLocatorInterface $serviceLocator)
     {
-        return new ModuleOptions($serviceLocator->get('Config')['zfc_rbac']);
+        /* @var \Rbac\Rbac $rbac */
+        $rbac = $serviceLocator->get('Rbac\Rbac');
+
+        /* @var \ZfcRbac\Service\RoleService $roleService */
+        $roleService = $serviceLocator->get('ZfcRbac\Service\RoleService');
+
+        /* @var \ZfcRbac\Assertion\AssertionPluginManager $assertionPluginManager */
+        $assertionPluginManager = $serviceLocator->get('ZfcRbac\Assertion\AssertionPluginManager');
+
+        /* @var \ZfcRbac\Options\ModuleOptions $moduleOptions */
+        $moduleOptions = $serviceLocator->get('ZfcRbac\Options\ModuleOptions');
+
+        $authorizationService = new AuthorizationService($rbac, $roleService, $assertionPluginManager);
+        $authorizationService->setAssertions($moduleOptions->getAssertionMap());
+
+        return $authorizationService;
     }
 }
