--- src/Factory/AuthenticationIdentityProviderFactory.php
@@ -1,44 +1,0 @@
-<?php
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license.
- */
-
-namespace ZfcRbac\Factory;
-
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
-use ZfcRbac\Identity\AuthenticationIdentityProvider;
-
-/**
- * Factory to create the authentication identity provider
- *
- * @author  Michaël Gallego <mic.gallego@gmail.com>
- * @licence MIT
- */
-class AuthenticationIdentityProviderFactory implements FactoryInterface
-{
-    /**
-     * {@inheritDoc}
-     * @return AuthenticationIdentityProvider
-     */
-    public function createService(ServiceLocatorInterface $serviceLocator)
-    {
-        /* @var \Zend\Authentication\AuthenticationService $authenticationProvider */
-        $authenticationProvider = $serviceLocator->get('Zend\Authentication\AuthenticationService');
-
-        return new AuthenticationIdentityProvider($authenticationProvider);
-    }
-}
